fixed bug in scan_red with float voxel sizes
git-svn-id: svn://svn.code.sf.net/p/slam6d/code/trunk@803 51f1e6e7-7289-49ad-8748-f09e43fc6daa

+++ src/slam6d/scan_red.cc
@@ -339,7 +339,7 @@ void scan2mat(Scan *source, cv::Mat &mat)
 }
 
 void reduce_octree(Scan *scan, vector<cv::Vec4f> &reduced_points, int octree,
-                   int red, bool use_reflectance)
+                   double red, bool use_reflectance)
 {
   if (use_reflectance) {
     unsigned int types = PointType::USE_REFLECTANCE;
